chore(ComponentExample): extract out vanilla instantiation code
This change extracts out instantiation/release code of vanilla
JavaScript library from `<ComponentExample>` and plain-HTML-specific
code from `<CodeExample>`. Doing so paves the cowpath toward better
integration of React live demo to those component.

Refs carbon-design-system/carbon#1936.
