+++ src/components/CodeExample/CodeExample.js
@@ -3,7 +3,11 @@ import PropTypes from 'prop-types';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import classnames from 'classnames';
 import ReactGA from 'react-ga';
-import Prism from 'prismjs';
+import { highlightAll } from 'prismjs/components/prism-core';
+import 'prismjs/components/prism-clike';
+import 'prismjs/components/prism-javascript';
+import 'prismjs/components/prism-markup';
+import 'prismjs/components/prism-jsx';
 import { Copy20, ChevronDown16 } from '@carbon/icons-react';
 
 import './prism.css';
@@ -11,6 +15,11 @@ import './prism.css';
 class CodeExample extends Component {
   static propTypes = {
     htmlFile: PropTypes.string,
+    codeClassName: PropTypes.string,
+  };
+
+  static defaultProps = {
+    codeClassName: 'language-html',
   };
 
   state = {
@@ -21,12 +30,12 @@ class CodeExample extends Component {
 
   componentDidMount = () => {
     if (this.codeBlock.offsetHeight > 190) {
-      this.setState({ showBtn: true });
+      this.setState({ showBtn: true }); // eslint-disable-line react/no-did-mount-set-state
     }
   };
 
   componentDidUpdate = () => {
-    Prism.highlightAll();
+    highlightAll();
   };
 
   handleCopy = () => {
@@ -48,7 +57,7 @@ class CodeExample extends Component {
   };
 
   render() {
-    const { htmlFile } = this.props;
+    const { htmlFile, codeClassName } = this.props;
     const copyBtnClass = classnames({
       'bx--btn--copy__feedback': true,
       'bx--btn--copy__feedback--displayed': this.state.copied,
@@ -81,7 +90,7 @@ class CodeExample extends Component {
           }}>
           <span className="code-title">Code:</span>
           <pre className="line-numbers">
-            <code className="language-html">{htmlFile}</code>
+            <code className={codeClassName}>{htmlFile}</code>
           </pre>
         </div>
         <CopyToClipboard text={htmlFile} onCopy={this.handleCopy}>
